--- .spotify_creds.json
@@ -1,1 +1,0 @@
-{"client_id": "f2292fcf102b48a29714a1ae913589e8", "client_secret": "a492f50e54c8487a8ed5c872745d1486"}
