+++ src/tracktidy/__init__.py
@@ -1,0 +1,5 @@
+"""
+TrackTidy - Music Manager for DJs and Music Enthusiasts
+"""
+
+__version__ = "0.1.0"
