+++ src/tracktidy/batch/__init__.py
@@ -1,0 +1,3 @@
+"""
+Batch processing functionality for TrackTidy
+"""
