+++ src/tracktidy/ui/__init__.py
@@ -1,0 +1,3 @@
+"""
+User interface components for TrackTidy
+"""
